rename timeslot to workerTimeslot for greater clarity as per amy's comment
