+++ server/models/handlers/workerTimeslot.js
@@ -1,4 +1,4 @@
-const timeslotModel = require('../db/timeslot');
+const timeslotModel = require('../db/workerTimeslot');
 
 // Adds a timeslot into the database and returns true upon successful completion. 
 async function addWorkerTimeslot( slotId, schoolId, workerId, status, date) {
